+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Disable.java
@@ -38,17 +38,17 @@ public class Disable implements ICommand<MessageEvent<Bot>>
 		}
 		else 
 		{
-			if(!CMDListener.channelStates.containsKey(args[1]))
+			if(!CMDListener.channelStates.containsKey(args[1].replace("#", "")))
 			{
-				CMDListener.channelStates.put(args[1], false);
+				CMDListener.channelStates.put(args[1].replace("#", ""), false);
 				Utilities.chanMsg(event, L10N.getString("disable.success", event));
 				Logging.info("Disabled bot in " + args[1]);
 			}
 			else
 			{
-				if(CMDListener.channelStates.get(args[1]))
+				if(CMDListener.channelStates.get(args[1].replace("#", "")))
 				{
-					CMDListener.channelStates.put(args[1], false);
+					CMDListener.channelStates.put(args[1].replace("#", ""), false);
 					Core.bot.sendCustomMessage(args[1], L10N.getString("disable.success", event));
 					Logging.info("Disabled bot in " + args[1]);
 				}